pyproject.toml
[tool.poetry]
name = "fyp-server"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
cryptography = "^44.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.5"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
